py-modules/cli/pyproject.toml
[project]
name = "macrostrat.cli"
version = "3.0.0"
description = "Macrostrat's command-line interface"
authors = [
    { name = "Daven Quinn", email = "dev@davenquinn.com" },
    { name = "Casey Idzikowski" },
    { name = "John J Czaplewski", email = "jczaplewski@wisc.edu" },
]
requires-python = ">=3.12,<4"
readme = "README.md"
dependencies = [
    "Fiona>=1.8.22,<2",
    "PyMySQL>=1.0.2,<2",
    "PyYAML~=6.0",
    "SQLAlchemy>=2.0.0,<3",
    "dynaconf>=3.1.12,<4",
    "geopandas>=0.14.1,<0.15",
    "ipython>=8.5.0,<9",
    "macrostrat.app-frame>=2.2.3,<3",
    "macrostrat.database>=3.5.4,<4",
    "macrostrat.utils>=1.3.0,<2",
    "macrostrat.dinosaur>=3.0.1,<4",
    "macrostrat.core",
    "macrostrat.integrations",
    "macrostrat.map-integration",
    "macrostrat.match-utils",
    "criticalmaas.ta1-geopackage>=0.2.0,<0.3",
    "mapboard.topology-manager",
    "htpheno.radosgw-admin-client",
    "numpy>=1.23.4,<2",
    "psycopg2-binary>=2.9.4,<3",
    "pyproj>=3.4.0,<4",
    "python-dotenv>=1.0.0,<2",
    "requests>=2.28.1,<3",
    "rich>=13,<14",
    "scipy>=1.9.3,<2",
    "shapely>=2.0.1,<3",
    "spectra>=0.0.11,<0.0.12",
    "tiletanic>=1.1.0,<2",
    "tqdm>=4.65.0,<5",
    "typer>=0.12,<0.13",
    "click>=8.1.0,<8.2.0",
    "docker>=7,<8",
    "toml>=0.10.2,<0.11",
    "greenlet>=3.0.3,<4",
    "paramiko>=3.4.0,<4",
    "pytest>=8.3.4,<9",
    "setuptools>=75.8.2,<81",
    "openpyxl>=3.1.5,<4",
    "results==1.4.0",
]

[project.scripts]
macrostrat = "macrostrat.cli.entrypoint:main"

[project.optional-dependencies]
gis = [
    "macrostrat.map-integration[gis]"
]

[dependency-groups]
dev = [
    "snakeviz>=2.2.0,<3",
    "macrostrat.package_tools>=1.1.0,<2",
]

[tool.uv]
default-groups = [
    "dev",
]

[tool.uv.sources]
"macrostrat.map-integration" = { path = "../map-integration", editable = true }
"macrostrat.core" = { path = "../core", editable = true }
"macrostrat.integrations" = { path = "../integrations", editable = true }
"mapboard.topology-manager" = { path = "../../submodules/topology-manager", editable = true }
"htpheno.radosgw-admin-client" = { path = "../../submodules/storage-admin", editable = true }
"macrostrat.match-utils" = { path = "../match-utils", editable = true }

[tool.hatch.build.targets.sdist]
include = ["macrostrat"]

[tool.hatch.build.targets.wheel]
include = ["macrostrat"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
